requirements.txt
streamlit>=1.30.0
openai>=1.0.0
tavily-python>=0.3.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
